services/booking-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-firestore==2.11.1
flask==3.0.3
